requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
pyodbc==5.2.0
PySocks==1.7.1
requests==2.32.3
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0